requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.3
django-admin==2.0.1
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-heroku==0.3.1
django-registration==3.1.1
django-six==1.0.4
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
pyuploadcare==2.7.0
requests==2.25.0
screen==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
xlwt==1.3.0
